+++ biz.aQute.bnd.reporter/test/biz/aQute/bnd/reporter/codesnippet/JavaSnippetReaderTest.java
@@ -5,6 +5,7 @@ import java.io.FileNotFoundException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
 import junit.framework.TestCase;
 
 public class JavaSnippetReaderTest extends TestCase {
@@ -174,7 +175,8 @@ public class JavaSnippetReaderTest extends TestCase {
     final Map<String, Integer> idCache = new HashMap<>();
     final JavaSnippetReader r = new JavaSnippetReader();
     r.init((i) -> {
-      final Integer count = idCache.put(i, idCache.getOrDefault(i, 0) + 1);
+			final Integer count = idCache.put(i,
+					Integer.valueOf(idCache.getOrDefault(i, Integer.valueOf(0)).intValue() + 1));
       if (count != null) {
         return i + count;
       } else {
